Fixed -Wsign-compare warnings in JRE build. 	Change on 2014/06/18 by tball <tball@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=69493943

+++ jre_emul/Classes/IOSArray.m
@@ -113,7 +113,7 @@ void IOSArray_throwOutOfBounds(NSUInteger size, NSUInteger index) {
     return @"[]";
   }
   NSString *result = @"[";
-  for (int i = 0; i < size_; i++) {
+  for (NSUInteger i = 0; i < size_; i++) {
     NSString *separator = i < size_ - 1 ? @", " : @"]";
     result = [result stringByAppendingFormat:@"%@%@",
                      [self descriptionOfElementAtIndex:i], separator];
